// src/types/Deal.ts
export interface Deal {
    id?: string;
    user_id?: string;
    business_name?: string;
    asking_price?: number;
    annual_revenue?: number;
    annual_net_income?: number;
    ffe_included?: boolean;
    inventory_included?: boolean;
    real_estate_included?: boolean;
    ffe_value?: number;
    inventory_value?: number;
    purchasing_real_estate?: boolean;
    real_estate_value?: number;
    annual_rent?: number;
    buyer_min_salary?: number;
    working_capital_requirement?: number;
    capex_maintanence?: number;
    capex_new_investments?: number;
    due_to_seller_business?: number;
    due_to_seller_real_estate?: number;
    total_due_to_seller?: number;
    total_cash_at_closing_to_seller?: number;
    seller_financing_paid_to_seller?: number;
    working_capital_required?: number;
    loan_closing_costs?: number;
    total_use_of_funds?: number;
    buyer_cash?: number;
    buyer_cash_percent?: number;
    seller_financing?: number;
    seller_financing_percent?: number;
    term_loan?: number;
    loan_closing_costs_percent?: number;
    loan_payments?: number;
    equipment_assets?: number;
    inventory?: number;
    working_capital?: number;
    loan_amount?: number;
    loan_closing_cost?: number;
    ebitda?: number;
    loan_term?: number;
    interest_rate?: number;
}